Use the right python to run playwright with uv (#11738)
## Describe your changes

This prevents an error where playwright is not detected as installed,
since we just uv installed it but then regular invoke it!

## GitHub Issue Link (if applicable)

## Testing Plan

I did not write any additional tests for this, but manually tested it.
If it breaks the current building path (which presumably must be
different than how I installed it) then the current CI will fail.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.