[fix] Resolving some regressions introduced via top nav (#11732)
## Describe your changes

Two main changes/bugs were fixed:
1. There was a race condition when apps were deployed such that
window.innerWidth was coming back as 0, satisfying the `<
breakpoints.md` condition. This was remedied with `useLayoutEffect`.
This fixes https://github.com/streamlit/streamlit/issues/11710.
2. CSS unhappiness where it was growing until its `maxWidth`. This was
fixed by adding a `minWidth: 0` to the appropriate spot. I also cleaned
up some styles while touching those components. This fixes
https://github.com/streamlit/streamlit/issues/11702


When I was creating an exhaustive test suite (`st_main_layout_test.py`)
I realized there was actually another bug present. There was a missing
case in the logic to account for the `EXPANDED` state for
`initial_sidebar_state`. I then discovered that there already existed
this helper in utils, with a lovely switch case that handled all of
these. So I utilized that and this is now resolved.


## GitHub Issue Link (if applicable)

Closes https://github.com/streamlit/streamlit/issues/11702
Closes https://github.com/streamlit/streamlit/issues/11710

## Testing Plan

I added e2e tests separate from the sidebar tests that focus
specifically on the main layout of the app including the initial sidebar
state, and various viewports, and uses snapshots to ensure content is
not overflowing off the page as was shown in
https://github.com/streamlit/streamlit/issues/11702. I think there's
opportunity to better consolidate these tests, but this seems sufficient
for now.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.

---------

Co-authored-by: Bob Nisco <bob.nisco@snowflake.com>