Fix the `websocket_ping_interval` to be at least `websocket_ping_timeout` (#11693)
## Describe your changes

Tornado changed the logic on how `websocket_ping_interval` and
`websocket_ping_timeout` are handled. With the latest version, the
`websocket_ping_interval` can never be less than the
`websocket_ping_timeout`. This leads to an unexpected ping timeout of 1s
when the current Streamlit is using the latest tornado version. To fix
this, we are updating the `websocket_ping_interval` to use 30s
(suggestion from Tornado maintainer
https://github.com/tornadoweb/tornado/pull/3376#issuecomment-2987096711)

This could potentially bring back an old issue with dropping connections
in specific proxy setups -
https://github.com/streamlit/streamlit/issues/3196. Something we need to
watch out for after this gets released.

## GitHub Issue Link (if applicable)

- Closes https://github.com/streamlit/streamlit/issues/11670

## Testing Plan

- Added unit test.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.