Remove python modules discovery from makefile (#11757)
## Describe your changes

The pytest commands in the makefile used `PYTHON_MODULES` as part of the
run command. But that is unnecessary since 1) there aren't any tests in
our main module and 2) we already provide the path to the tests folder
in the command call.

## GitHub Issue Link (if applicable)

- Closes https://github.com/streamlit/streamlit/issues/11741

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.