[fix] Fixing the sidebar font weight and spacing for sections (#11814)
## Describe your changes

<!-- If it's a visual change, please include a screenshot or video! -->

Two changes:
1) font weight of the section headers
2) spacing between the nav sections in sidebar


| Before | After |
|--------|-------|
|
![Before](https://github.com/user-attachments/assets/67a805ba-0d04-4ae5-9078-977c822f69cc)
|
![After](https://github.com/user-attachments/assets/2b5da6de-9fdd-4b8c-a541-fa98341bddfa)
 


## GitHub Issue Link (if applicable)

No formal issue link. The padding was requested in the design mocks for
top nav and scoped out as a follow up. The font weight brought up in the
original pr here:
https://github.com/streamlit/streamlit/pull/11349#discussion_r2146340177

## Testing Plan

Changes captured in the existing snapshots.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.