Fix protobuf generation quoting error on windows (#11776)
## Describe your changes.
This allows the protobuf step to be run even if the filepath has special
characters in it. ...Sort of. However, there is a bug in pbts that also
needs to be fixed before that works.

## GitHub Issue Link (if applicable)

Partially fixes https://github.com/streamlit/streamlit/issues/11742

## Testing Plan

I have manually tested this, and if it breaks the build, the current CI
will catch it. Unless the current CI doesn't rely on the protobufs for
some reason? Honestly I'm not sure what this step is used for, I just
know that it runs during the make steps.

Could be cool to write a test for this in the CI, but I'm not sure it's
worth it.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.

---------

Co-authored-by: lukasmasuch <lukas.masuch@gmail.com>