Make dataframe error indicator adapt to sizing from theme (#11682)
## Describe your changes

Mini change to make the warning indicator in dataframe not be based on
static pixel value and instead just fill the padding.

Before:


![image](https://github.com/user-attachments/assets/b8905f04-7118-49c5-8aa6-fc14ab0b4309)

After:


![image](https://github.com/user-attachments/assets/cef48f29-62c6-46a9-b95f-f6a91139967d)


## Testing Plan

- Existing tests should cover this well so that it doesn't change with
the normal base font size. Testing this with different base font sizes
feels be overkill.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.

---------

Co-authored-by: Laura Wilby <laura.wilby@snowflake.com>