[fix] No more sidebar flickering when initial_sidebar_state="collapsed" (#11861)
## Describe your changes

This change fixes the bug linked below and adds a bunch of testing to
ensure this doesn't happen again without us catching it. Additionally,
it was discovered that there were some tests that were locking in this
buggy behavior, so those have been updated as well.

<!-- If it's a visual change, please include a screenshot or video! -->

## GitHub Issue Link (if applicable)

This PR fixes #11848 

## Testing Plan

Added a new e2e test file as I felt it was merited to ensure this
functionality doesn't regress. The test injects some javascript to look
for any transient flickers that might occur that could be missed by
playwright.


---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.