[Fix] Flicker on facet altair charts. (#11833)
## Describe your changes

There are a few issues with facet charts that this PR fixes:

- Autosize does not work with facet charts. Currently, it seems we are
mainly seeing a warning and it is not directly causing the flicker, but
this is a source of chart issues and it's better to remove it.
- Using `fit-content` on the parent containers of the facet chart seems
to cause a loop where the spec is altered leading to a loop. It seems
the trigger for this was the change to use the `width` measured from
`StyledVegaLiteChartContainer` instead of from `StyledFullScreenFrame`
introduced
[here](https://github.com/streamlit/streamlit/commit/50c7ecc633c58059b7f4e49088c2a42a9de37f04#diff-92f0e57b8f2d70900bd5b548ed97c7fb579f2880d48faaea8b7cfac004d479c0).
Although the `width` is only added to the spec when the chart is full
screen (which is a trigger for flickering), this width is also part of
the dependency array. So it seems there is a loop where because the
parent container's width is dependent on the chart, the chart must also
be dependent on the parent width in the vega code (we are not using it
outside of full screen) and we enter a loop. This seems only applicable
to the facet charts, so there must be something different about how vega
handles dimensions with these charts leading to the loop.

<!-- If it's a visual change, please include a screenshot or video! -->

## GitHub Issue Link (if applicable)

## Testing Plan

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.