Fix issue with missing fragment after reconnect (#11890)
## Describe your changes

Ensure that we request a full rerun if the websocket connection
reconnects and if the previous script run contained fragments. This is
to ensure that the app session has initialized all of the relevant app
fragments.

## GitHub Issue Link (if applicable)

- Closes https://github.com/streamlit/streamlit/issues/11660


## Testing Plan

- Added unit tests.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.