Make crossorigin property configurable for media elements (#11948)
## Describe your changes

<!-- If it's a visual change, please include a screenshot or video! -->

This PR adds a new field to the `/host-config` endpoint:
`resourceCrossOriginMode `
If this field is set to `"anonymous"`, the `crossOrigin="anonymous"`
(see
https://developer.mozilla.org/en-US/docs/Web/API/HTMLImageElement/crossOrigin)
is added to the media elements `img`, `audio`, and `video` including
when rendered via markdown.
This allows those elements to render when loaded from a cross origin
target, though the browser does not add any credentials automatically to
the request.
If the value is undefined, the `crossOrigin` property is not set at all.

For testing reasons, the server can also send the
`setAnonymousCrossOriginPropertyOnMediaElements: true` field. This field
can be removed in a few weeks.

## GitHub Issue Link (if applicable)

## Testing Plan

- Explanation of why no additional tests are needed
- Unit Tests (JS and/or Python)
- Added unit tests to ensure that the attribute `crossOrigin` is set
based on the new lib context
- E2E Tests
- Any manual testing needed?

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.