Hide quickselect in `date_input` if min date is more than 2 years ago (#11999)
## Describe your changes

Quickselect in date input is allowing to go back to up to 2 years. In
order to comply with the configured min date, we are only activating
quickselect if the min date is at least 2 years ago

## GitHub Issue Link (if applicable)

- Closes https://github.com/streamlit/streamlit/issues/11939

## Testing Plan

- Added unit test.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.