[fix] Sidebar toggle state reset on page resize.  (#12024)
## Describe your changes

This PR persists the sidebar toggle state in `localStorage` so that it
can be taken into consideration when calculating whether the sidebar is
collapsed or not. This fixes and issue where resizing the page was
causing the sidebar to revert to its config state regardless of whether
it had been toggled open/closed.

This also implements this enhancement request #11709

Tests for width state are also added since we didn't have specific tests
for that already.

<!-- If it's a visual change, please include a screenshot or video! -->

## GitHub Issue Link (if applicable)

#12016


## Testing Plan

- Unit Tests (JS and/or Python) ✅ 
- E2E Tests ✅ 
- Manually tested ✅  (didn't test with page_config)

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.