Use crossOrigin property only when making requests to the Streamlit backend (#12054)
## Describe your changes

Follow-up of https://github.com/streamlit/streamlit/pull/11948
In https://github.com/streamlit/streamlit/pull/11948, the option was
introduced to set the `crossorigin` attribute of media elements.
However, the attribute should only be set when the asset is loaded from
the Streamlit backend server when configured by setting the
`window.__streamlit?.BACKEND_BASE_URL` variable, but not if it is loaded
from *external* servers. Setting it always can lead to issues if the
server does not respond with the appropriate headers.
When configuring the `window.__streamlit?.BACKEND_BASE_URL` variable,
the request can go to a different origin than from where the Streamlit
assets are loaded (in case of an iframe). If
`window.__streamlit?.BACKEND_BASE_URL` is set and would be the same
origin as where the iframe is coming from, the `crossorigin` attribute
would be ignored by the browser as it is a same-origin request in this
case.

## GitHub Issue Link (if applicable)

## Testing Plan

- Explanation of why no additional tests are needed
- Unit Tests (JS and/or Python)
  - Added unit tests for the new utils function
- E2E Tests
- Any manual testing needed?
  - Tested manually locally

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.