Log warning instead of assert when config file is present (#12093)
## Describe your changes

This assert can cause trouble when running/discovering tests locally
(e.g. via test explorer) and if there is a config file present somewhere
in the path. This changes it from an assert to a warning log.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.