[feat] Add JSON5 validation in pre-commit (#12114)
## Describe your changes

- Follow up from https://github.com/streamlit/streamlit/pull/11580
- There are some JSON files that are allowed to have comments in them
(VSCode settings, tsconfig)
- Rather than skipping them, we can parse them with JSON5 to get a
stronger sense that they are well-formatted
- _Technically speaking_ JSON5 is a superset of JSON with comments. I
was unable to find a solid JSON with comments parser, JSON5 is more well
supported in the ecosystem. I figure this is a fine tradeoff as they
effectively are interchangeable for our purposes.

## GitHub Issue Link (if applicable)

## Testing Plan

- Explanation of why no additional tests are needed
- Unit Tests (JS and/or Python)
- E2E Tests
- Any manual testing needed?

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.