[fix] st.chat_input collapses after submit (#12132)
## Describe your changes

This PR addresses a regression where the logic for resetting the height
of `st.chat_input` after submission was broken by some refactoring. It
re-introduces the code to reset the scroll height to 0 after submit. The
PR also adds an e2e test for the user interaction.

<!-- If it's a visual change, please include a screenshot or video! -->

## GitHub Issue Link (if applicable)

Fixes #12079

## Testing Plan

- E2E Tests ✅ 

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.