[fix] iframe and html change in default width when no width specified. (#12148)
## Describe your changes

When adding `horizontal_alignment` on `st.container`, the default for
`align-items` on `StyledFlexContainerBlock` was specified as `start`.
This is consistent with the default we want for the containers, but this
also changed the wrapper for the main app contents. This changed the
default width of `st.iframe`, `st.html` and also the loading skeletons
(fixed in previous PR). I am changing this to `stretch` out of caution.
The default for `st.container` is already defined in the python code.

To ensure `st.iframe` and `st.html` work correctly, this PR also updates
`iframe` and `html` to use the new fields on `Element.proto` consistent
with other elements.

<img width="840" height="475" alt="Screenshot 2025-08-07 at 3 53 13 PM"
src="https://github.com/user-attachments/assets/26f16c2f-0f26-4700-b9ba-407dd2457f9c"
/>

<!-- If it's a visual change, please include a screenshot or video! -->

## GitHub Issue Link (if applicable)

## Testing Plan

- Unit Tests (JS and/or Python) ✅ 
- E2E Tests ✅ 

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.

---------

Co-authored-by: graphite-app[bot] <96075541+graphite-app[bot]@users.noreply.github.com>