[fix] Cache layout_config data and use during replay (#12183)
## Describe your changes

While working on width for `st.image` I discovered that since we only
cache the inner element proto (e.g. Image.proto, Code.proto), when an
element with width/height added in the new style on Element.proto is
cached it does not replay with the correct width/height.

This PR caches the layout_config as well so that it can be replayed as
well.

<!-- If it's a visual change, please include a screenshot or video! -->

## GitHub Issue Link (if applicable)

## Testing Plan

- Unit Tests (JS and/or Python) ✅ 
- E2E Tests ✅ 

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.