Update credentials.py: try to fix showEmailPrompt in case where there is no email (#12202)
## Describe your changes

Fixes #12166 “If showEmailPrompt is false, but there is no email
configured, "Activation email not valid" error results”

## GitHub Issue Link (if applicable)

https://github.com/streamlit/streamlit/issues/12166

## Testing Plan

This should be testable, and in fact the tests that were written for the
original PR just have a small blindspot that needs to be fixed, for this
case. Perhaps this addresses it:

This PR changes
`test_streamlit_folder_not_created_when_show_email_prompt_false` to
expect a succeeding error code instead of a error code — it's surprising
to me that it ever expected a failing error code, and maybe that was a
typo that caused this undesired behavior to slip through (or maybe I
don't understand what it was doing).

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.