[AdvancedLayouts] Updates to st.image width/height and adding width/height to st.pyplot.  (#11952)
## Describe your changes

This PR is part of the [AdvancedLayouts] project and updates `width` on
`st.image` to use the new width style and adds `width` to `st.pyplot`.

This PR also begins the process of deprecating use_container_width. The
default is updated to None, and width will be used instead. If the user
explicitly passes a value for use_container_width then that will take
precedence. use_container_width=True is equivalent to width="stretch"
and use_container_width=False is equivalent to width="content". The user
will be given a warning and suggestion to use width instead. We will
remove use_container_width after 12-31-2025`.

<!-- If it's a visual change, please include a screenshot or video! -->

## GitHub Issue Link (if applicable)

## Testing Plan

- Unit Tests (JS and/or Python) ✅ 
- E2E Tests ✅ 

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.

---------

Co-authored-by: Bob Nisco <bob.nisco@snowflake.com>